Renames property to properties.
