+++ ui/revenuecatui/src/main/kotlin/com/revenuecat/purchases/ui/revenuecatui/components/modifier/Background.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.Stable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.graphics.Shape
-import com.revenuecat.purchases.ui.revenuecatui.components.property.ColorStyle
+import com.revenuecat.purchases.ui.revenuecatui.components.properties.ColorStyle
 
 @JvmSynthetic
 @Stable
