+++ ui/revenuecatui/src/main/kotlin/com/revenuecat/purchases/ui/revenuecatui/components/properties/ShadowStyle.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.ui.revenuecatui.components.property
+package com.revenuecat.purchases.ui.revenuecatui.components.properties
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Immutable
